requirements.txt
requests==2.22.0
transformers==4.10.1
flask==1.1.2
tensorflow==2.7 
pillow==9.0